Login and register models moved to Viwmodels folder
